Polish : Burrow Sense & Photosensitive
Improved visually by "White" outline. Highlighting the blocks where the ore is. Additionally when the ore found is destroyed or mined the highlight will be gone and the cooldown of the highlight is 10 seconds.

For the photosensitive, remove the fire effect and instead add the fire particle that shows damage when in sunlight.